AArch64: Migrate calls and returns to ISLE.
